Updates planar gripper sim parameters to match hardware. Also, moved BrickFace and Finger enums to planar_gripper_common.
